fr: Begin with a GRE/FrameRelay test due the recent regression
The framerelay code is seldomly used and the socket clean ups introduced
a regression. Create a testcase that will work as a user not having the
right capabilities to create raw sockets. This is done with dlopen/dlsym
to convert the socket raw to something else.

This is failing right now and needs to amend it once the fix is in place.