gb: Fix gprs_ns_rx_reset to not create NS-VC duplicates
Under special circumstances (see below) receiving a NS-RESET leads to
duplicated NS-VC entries.

This patch changes gprs_ns_rx_reset() to check for this case and to
eventually delete the old entry. A new counter NS_CTR_REPLACED is
incremented each time when the NS-VC object is replaced. A new signal
S_NS_REPLACED is added which gets dispatched in this case, too.

This happens when the source port of a NS-VC changes to a new one
that has already been used by another NS-VC.

Ticket: OW#874
Sponsored-by: On-Waves ehf